Created AtlanObservability class
- Created a Observability superclass which would be inherited by all the Adaptors. It will have function to push to objectStore, log retention capabilities etc.
