+++ tests/unit/common/test_logger_adapter.py
@@ -266,70 +266,66 @@ def mock_parquet_file(tmp_path):
 
 
 @pytest.fixture(autouse=True)
-def clear_log_buffer():
+def clear_log_buffer(logger_adapter):
     """Clear the log buffer before each test."""
-    AtlanLoggerAdapter._log_buffer.clear()
+    logger_adapter._log_buffer.clear()
     yield
-    AtlanLoggerAdapter._log_buffer.clear()
+    logger_adapter._log_buffer.clear()
 
 
 @pytest.mark.asyncio
 async def test_parquet_sink_buffering(mock_parquet_file):
     """Test that parquet_sink properly buffers logs."""
     with create_logger_adapter() as logger_adapter:
-        # Mock the parquet file path
-        with mock.patch(
-            "application_sdk.common.logger_adaptors.PARQUET_PATH",
-            str(mock_parquet_file),
-        ):
-            # Create a test message
-            test_message = mock.MagicMock()
-            level_mock = mock.MagicMock()
-            level_mock.name = "INFO"  # Set the name attribute directly
-            test_message.record = {
-                "time": datetime.now(),
-                "level": level_mock,
-                "extra": {"logger_name": "test_logger"},
-                "message": "Test message",
-                "file": mock.MagicMock(path="test.py"),
-                "line": 1,
-                "function": "test_function",
-            }
+        # Set the parquet file path directly on the instance
+        logger_adapter.parquet_path = str(mock_parquet_file)
 
-            # Call parquet_sink
-            await logger_adapter.parquet_sink(test_message)
+        # Create a test message
+        test_message = mock.MagicMock()
+        level_mock = mock.MagicMock()
+        level_mock.name = "INFO"  # Set the name attribute directly
+        test_message.record = {
+            "time": datetime.now(),
+            "level": level_mock,
+            "extra": {"logger_name": "test_logger"},
+            "message": "Test message",
+            "file": mock.MagicMock(path="test.py"),
+            "line": 1,
+            "function": "test_function",
+        }
 
-            # Verify log was added to buffer
-            assert len(logger_adapter._log_buffer) == 1
-            buffered_log = logger_adapter._log_buffer[0]
-            assert buffered_log["message"] == "Test message"
-            assert buffered_log["level"] == "INFO"
-            assert buffered_log["logger_name"] == "test_logger"
+        # Call parquet_sink
+        await logger_adapter.parquet_sink(test_message)
+
+        # Verify log was added to buffer
+        assert len(logger_adapter._log_buffer) == 1
+        buffered_log = logger_adapter._log_buffer[0]
+        assert buffered_log["message"] == "Test message"
+        assert buffered_log["level"] == "INFO"
+        assert buffered_log["logger_name"] == "test_logger"
 
 
 @pytest.mark.asyncio
 async def test_parquet_sink_error_handling(mock_parquet_file):
     """Test that parquet_sink handles errors gracefully."""
     with create_logger_adapter() as logger_adapter:
-        # Mock the parquet file path
-        with mock.patch(
-            "application_sdk.common.logger_adaptors.PARQUET_PATH",
-            str(mock_parquet_file),
-        ):
-            # Create a test message with invalid data
-            test_message = mock.MagicMock()
-            test_message.record = {
-                "time": datetime.now(),
-                "level": mock.MagicMock(name="INFO"),
-                "extra": {"logger_name": "test_logger"},
-                "message": "Test message",
-                "file": None,  # This will cause an error
-                "line": 1,
-                "function": "test_function",
-            }
+        # Set the parquet file path directly on the instance
+        logger_adapter.parquet_path = str(mock_parquet_file)
 
-            # Call parquet_sink - should not raise exception
-            await logger_adapter.parquet_sink(test_message)
+        # Create a test message with invalid data
+        test_message = mock.MagicMock()
+        test_message.record = {
+            "time": datetime.now(),
+            "level": mock.MagicMock(name="INFO"),
+            "extra": {"logger_name": "test_logger"},
+            "message": "Test message",
+            "file": None,  # This will cause an error
+            "line": 1,
+            "function": "test_function",
+        }
 
-            # Verify buffer is empty (error was handled)
-            assert len(logger_adapter._log_buffer) == 0
+        # Call parquet_sink - should not raise exception
+        await logger_adapter.parquet_sink(test_message)
+
+        # Verify buffer is empty (error was handled)
+        assert len(logger_adapter._log_buffer) == 0
